[Refactor] 세션 Controller 외부,내부로 분리
